Make x axis ticks logic more flexible. Fixes #156 Fixes #118
We changed the logic of the x axis so that the user has more
control over which ticks will be visible on the axes and which
won't be.

It now will prioritise ticks with actual names over empty strings
and the overflow detection logic is less "agressive" allowing for
axis labels that are close to each other without reducing total
number of ticks too drastically.